+++ native/src/types/name.rs
@@ -134,8 +134,13 @@ impl Name {
     /// * `display` - A string containing the street name (Main St)
     ///
     /// ```
-    pub fn new(display: String, priority: i8, context: &Context) -> Self {
+    pub fn new(mut display: String, priority: i8, context: &Context) -> Self {
         let tokens = context.tokens.process(&display);
+
+        display = display
+            .replace(r#"""#, "")
+            .replace("\t", "")
+            .replace("\n", "");
 
         Name {
             display: display,